# study-service/src/main/resources/openapi/study-api.openapi.yml
openapi: 3.0.0
info:
  title: Study API
  version: 0.0.1
servers:
  - url: http://localhost:8083/api/study
paths:
...
  /schedule/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Schedule Information by Id
      tags:
        - Schedule
      operationId: getScheduleByGroupId
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Schedule by Id
      tags:
        - Schedule
      operationId: deleteSchedule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
...
  /group/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Group Information
      tags:
        - Group
      operationId: getGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Group By Id
      tags:
        - Group
      operationId: deleteGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
...
  /attendance-statistics:
    post:
      security:
        - bearerAuth: [ ]
      summary: Add/Update Statistics
      tags:
        - AttendanceStatistics
      operationId: addStatistics
      requestBody:
        description: Attendance Statistics Entity
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AttendanceStatisticsOpenApi'
      responses:
        '201':
          description: Added
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /attendance-statistics/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Attendance statistics Information
      tags:
        - AttendanceStatistics
      operationId: getStatisticsById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /attendance-statistics/group/{group_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Group's Attendance Statistics Information
      tags:
        - AttendanceStatistics
      operationId: getStatisticsByGroupId
      parameters:
        - in: path
          name: group_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /attendance-statistics/schedule/{scheduleId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get Attendance Statistics by Schedule Id
      tags:
        - AttendanceStatistics
      operationId: getStatisticsBySchedule
      parameters:
        - in: path
          name: scheduleId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /attendance-statistics/subject/{subject_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subject's Attendance Statistics Information
      tags:
        - AttendanceStatistics
      operationId: getStatisticsBySubjectId
      parameters:
        - in: path
          name: subject_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /subject/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subject Information
      tags:
        - Subject
      operationId: getSubjectById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Subject by Id
      tags:
        - Subject
      operationId: deleteSubject
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /download/homework/{fileId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Download File by It's Id
      tags:
        - File Download
      operationId: downloadHomeworkFileById
      parameters:
        - in: path
          name: fileId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /download/materials/{fileId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Download File by It's Id
      tags:
        - File Download
      operationId: downloadMaterialsFileById
      parameters:
        - in: path
          name: fileId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ScheduleOpenAPI:
      type: object
      required:
        - dateTime
        - studyGroupId
        - subjectId
        - classFormat
        - professorId
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        studyGroupId:
          type: integer
          format: int64
        subjectId:
          type: integer
          format: int64
        classFormat:
          $ref: "#/components/schemas/ClassFormat"
        auditorium:
          type: string
        linkForTheClass:
          type: string
        professorId:
          type: integer
          format: int64
    StudyGroupOpenAPI:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studentsIds:
          type: array
          description: studentsIDs
          items:
            type: integer
            format: int64
    SubjectOpenAPI:
      type: object
      required:
        - name
        - studyGroupsIds
        - creatorId
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studyGroupsIds:
          type: array
          items:
            type: integer
            format: int64
        studyMaterialsIds:
          type: array
          items:
            type: integer
            format: int64
        scheduleIds:
          type: array
          items:
            type: integer
            format: int64
        creatorId:
          type: integer
          format: int64
        editorsIds:
          type: array
          items:
            type: integer
            format: int64
    ClassFormat:
      type: string
      enum:
        - IN_PERSON
        - DISTANT
...
    AttendanceStatisticsOpenApi:
      type: object
      required:
        - scheduleId
        - userId
        - status
      properties:
        scheduleId:
           type: integer
           format: int64
        status:
          $ref: "#/components/schemas/AttendanceStatusOpenApi"
        studentId:
          type: integer
          format: int64
    AttendanceStatusOpenApi:
      type: string
      enum:
        - ATTENDED
        - MISSED
...